Return values of syscalls are back
* This is probably the last feature of ftrace broken by recent merger.

+++ frysk-core/frysk/proc/Syscall.java
@@ -218,7 +218,7 @@ public abstract class Syscall
 
 	case 'i':
 	    if ((int)retVal < 0)
-		return "-1 ERRNO=" + (-retVal);
+		return "-1 ERRNO=" + (-(int)retVal);
 
 	    // fall-through
 	default:
